fix: harden the registry against the ways it could be bricked or lied to
An audit of the base layer before the frontend goes on top of it.

registerRoot rejects the zero root, withdrawStatement refuses a second withdrawal,
and setNameGuard refuses an address without code — a guard that is an EOA or a
typo would make every registration revert, and only the owner could undo it.
ENSv2NameGuard checks the same for its registry, whose address is immutable.

segmentCount is now documented for what it is: a claim by the author. A root does
not reveal its leaf count, so the contract cannot check it, and a test pins that
behaviour down so nobody mistakes it for a check. The honest n comes from the
bundle, which verifyBundle holds against the root — nothing can be added, dropped
or moved there. Two tests cover that, and one records what the root does not cover:
the vectors, which hang on the CID alone.

The FFI bridge returns uint32 for the segment count, which removes the last three
forge lint warnings.

@xenova/transformers moves to devDependencies. It only runs the layer-3 measurement
scripts, and with it out of the runtime path npm audit --omit=dev is clean.

+++ script/js/tree.mjs
@@ -22,7 +22,7 @@ const { root, proofFor } = buildTree(segments);
 
 process.stdout.write(
   encodeAbiParameters(
-    [{ type: "bytes32" }, { type: "uint256" }, { type: "bytes32[]" }],
+    [{ type: "bytes32" }, { type: "uint32" }, { type: "bytes32[]" }],
     [root, BigInt(segments.length), proofFor(Number(rawIndex))],
   ),
 );
